Context and M2k: Drop the pimpl implementation and update all the classes accordingly.
