+++ include/libm2k/lidar.hpp
@@ -1,3 +1,24 @@
+/*
+ * Copyright (c) 2019 Analog Devices Inc.
+ *
+ * This file is part of libm2k
+ * (see http://www.github.com/analogdevicesinc/libm2k).
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU Lesser General Public License as published by
+ * the Free Software Foundation, either version 2.1 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU Lesser General Public License for more details.
+ *
+ * You should have received a copy of the GNU Lesser General Public License
+ * along with this program. If not, see <http://www.gnu.org/licenses/>.
+ *
+ */
+
 #ifndef LIDAR_H
 #define LIDAR_H
 
@@ -10,43 +31,31 @@
 using namespace std;
 
 namespace libm2k {
-	namespace contexts {
-		class LIBM2K_API Lidar : public Context {
-		public:
-			Lidar(std::string uri, iio_context* ctx, std::string name, bool sync);
-			virtual ~Lidar();
+namespace contexts {
+class LIBM2K_API Lidar : public virtual Context {
+public:
+	virtual ~Lidar() {}
 
-			void channelEnableDisable(string channel, bool enable);
-			map<string, vector<int8_t>> readChannels(unsigned int nb_samples);
-	   
-			void setApdBias(unsigned int bias);
-			void setTiltVoltage(unsigned int voltage);
-			void laserEnable();
-			void laserDisable();		
-			void setLaserPulseWidth(long long pulse_width);
-			void setLaserFrequency(long long frequency);
+	virtual void channelEnableDisable(string channel, bool enable) = 0;
+	virtual map<string, vector<int8_t>> readChannels(unsigned int nb_samples) = 0;
 
-			void enableChannelSequencer();
-			void disableChannelSequencer();
-			void setSequencerPulseDelay(long long ns);
+	virtual void setApdBias(unsigned int bias) = 0;
+	virtual void setTiltVoltage(unsigned int voltage) = 0;
+	virtual void laserEnable() = 0;
+	virtual void laserDisable() = 0;
+	virtual void setLaserPulseWidth(long long pulse_width) = 0;
+	virtual void setLaserFrequency(long long frequency) = 0;
 
-			void setChannelSequencerOpModeManual();
-			void setChannelSequencerOpModeAuto();
-			void setChannelSequencerOrderAutoMode(const char* order);
-			void setChannelSequencerOrderManualMode(const char* order);
-    
-			// void enableTempCompensation();
-			// void disableTempCompensation();
-			// int readSystemTemp();
-			// void setTempComputationCoeff(unsigned int);
-			// //// Req. Update the APD bias based on the system temperature.    
-			// void autoUpdateApdBias();
+	virtual void enableChannelSequencer() = 0;
+	virtual void disableChannelSequencer() = 0;
+	virtual void setSequencerPulseDelay(long long ns) = 0;
 
-		private:
-			class LidarImpl;
-			std::shared_ptr<LidarImpl> m_pimpl;
-		};
-	}
+	virtual void setChannelSequencerOpModeManual() = 0;
+	virtual void setChannelSequencerOpModeAuto() = 0;
+	virtual void setChannelSequencerOrderAutoMode(const char* order) = 0;
+	virtual void setChannelSequencerOrderManualMode(const char* order) = 0;
+};
+}
 }
 
 #endif
